+++ ceammc/ext/src/ui/ui_gain.cpp
@@ -30,6 +30,8 @@ UIGain::UIGain()
     , txt_min_(font_.font(), ColorRGBA::black(), ETEXT_DOWN_LEFT, ETEXT_JLEFT)
     , knob_pos_(0)
     , is_horizontal_(false)
+    , rel_mode_delta_(0)
+    , prop_relative_mode(0)
 {
 }
 
@@ -159,15 +161,32 @@ void UIGain::onBang()
 
 void UIGain::onMouseDown(t_object* view, const t_pt& pt, const t_pt& abs_pt, long modifiers)
 {
-    onMouseDrag(view, pt, modifiers);
+    const float new_value = (is_horizontal_) ? clip<float, 0, 1>(pt.x / width())
+                                             : clip<float, 0, 1>(1.0 - (pt.y / height()));
+
+    if (prop_relative_mode) {
+        rel_mode_delta_ = (new_value - knob_pos_);
+        return;
+    } else {
+        knob_pos_ = new_value;
+        redrawBGLayer();
+
+        if (prop_output_value)
+            onBang();
+    }
 }
 
 void UIGain::onMouseDrag(t_object* view, const t_pt& pt, long modifiers)
 {
-    if (is_horizontal_)
-        knob_pos_ = clip<float>(pt.x / width(), 0, 1);
+    // NOTE: no clip - it's done later
+    const float new_value = (is_horizontal_) ? pt.x / width()
+                                             : 1.0 - (pt.y / height());
+
+    // clip only here
+    if (prop_relative_mode)
+        knob_pos_ = clip<float, 0, 1>(new_value - rel_mode_delta_);
     else
-        knob_pos_ = clip<float>(1 - pt.y / height(), 0, 1);
+        knob_pos_ = clip<float, 0, 1>(new_value);
 
     redrawBGLayer();
 
@@ -242,6 +261,10 @@ void UIGain::m_dec()
 
 void UIGain::setup()
 {
+    static t_symbol* SYM_DB = gensym("db");
+    static t_symbol* SYM_MAX = gensym("max");
+    static t_symbol* SYM_MIN = gensym("min");
+
     UIDspFactory<UIGain> obj("ui.gain~");
     obj.addAlias("ui.hgain~");
     obj.addAlias("ui.vgain~");
@@ -249,13 +272,17 @@ void UIGain::setup()
     obj.addColorProperty("knob_color", _("Knob Color"), DEFAULT_ACTIVE_COLOR, &UIGain::prop_color_knob);
     obj.addHiddenFloatCbProperty("db", &UIGain::dbValue, &UIGain::setDbValue);
     obj.setPropertyDefaultValue("db", "-60");
+    obj.setPropertyUnits(SYM_DB, SYM_DB);
     obj.addHiddenFloatCbProperty("amp", &UIGain::ampValue, &UIGain::setAmpValue);
     obj.addIntProperty("max", _("Maximum value"), 0, &UIGain::prop_max, _("Bounds"));
     obj.addIntProperty("min", _("Minimum value"), -60, &UIGain::prop_min, _("Bounds"));
     obj.setPropertyRange("max", -12, 12);
     obj.setPropertyRange("min", -90, -30);
+    obj.setPropertyUnits(SYM_MAX, SYM_DB);
+    obj.setPropertyUnits(SYM_MIN, SYM_DB);
     obj.addBoolProperty("show_range", _("Show range"), true, &UIGain::prop_show_range, _("Misc"));
     obj.addBoolProperty("output_value", _("Output value"), false, &UIGain::prop_output_value, _("Main"));
+    obj.addBoolProperty("relative", _("Relative mode"), true, &UIGain::prop_relative_mode, _("Main"));
 
     obj.setDefaultSize(15, 120);
     obj.usePresets();